Refactor: Sposta logica Gestione Servizi GE e implementa logging modifiche
- Sposta le classi FilterManager, QueryBuilder, DataManager e le route
  relative alla Gestione Servizi GE da 'ordini_servizi.py' al nuovo
  file 'ordini_servizi_ge.py'.
- Aggiorna 'main.py' per includere il nuovo router 'servizi_ge_router'.
- Implementa il logging delle modifiche nella funzione
  'update_gestione_gs_data' in 'ordini_servizi_ge.py'.
  Le modifiche vengono registrate nella tabella 'carrefour_log',
  includendo utente, valore vecchio/nuovo, tabella, colonna e data.
- La tabella interessata dagli aggiornamenti e dal logging è
  'carrefour_contabilizzazione_originale'.
